Replace SharedSynchronizedApiClassesClassLoader by the default class loader
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This diff
We used a custom class loader that encapsulates both the `java.lang.ClassLoader#getPlatformClassLoade`r and the deprecated `javax.tools.ToolProvider.getSystemToolClassLoader()` likely for backward compatibility.

However, it looks like that this wrapper is no longer necessary, so I am trying to remove it.

Reviewed By: siaojiecai

Differential Revision: D77576914

fbshipit-source-id: 6d3574f9f581c37a03071e56ff2c14c557344596