Move ToolchainInfo out of prelude
Summary:
As Dmitry, we were leaking sabaton information into the open source buck2 distribution through the prelude.

Buck2 folks were nice enough to confirm we don't need to add the toolchain to the prelude for it to be usable in the way we need it so let's re-shuffle some definitions to remove it from the prelude.

Reviewed By: lunaticare

Differential Revision: D77903476

fbshipit-source-id: 38b44cf6d3122d88bc138567495708c6032a6d80