Remove exit point from CompilerDaemonRunner
Summary:
In case of failure, CompilerDaemonRunner exits directly.
This is a carry over from buck1.
We want the CompilerDaemonRunner to throw an exception and handle it on the call site, which is what this diff is about.
We also want to flush the stderr before exiting.

Meta:
More: https://fburl.com/workplace/lds4e4ic
In the long therm we should get rid of Isolated Steps completely: T230141909

Reviewed By: ywett02

Differential Revision: D77882581

fbshipit-source-id: b22018da6560b39adcfd6992b2438ef14ca2dc83