Make default true
Summary:
### Context
We are integrating the Kotlin Incremental Compiler (KSIC) into Buck to enable incremental compilation of Kotlin source files. This aims to enhance build speed by recompiling only modified source files within a target, rather than rebuilding the entire target.

Following a successful A/B test on large IG targets (those with 20 or more Kotlin source files), which showed a positive impact on both compilation time and build speed, we now want to experiment with making KSIC the default choice for IG. The goals of this testing phase are:
- To assess the overall effect of KSIC on IG
- To determine the threshold in terms of the number of Kotlin source files beyond which KSIC remains effective
- To establish whether sufficient experimental power can be achieved per app

### This diff
In D77595676 I switched default to false. However, it's likely better to keep it as true since that matches with expected behaviour of `incremental` flag.

Reviewed By: IanChilds

Differential Revision: D78145772

fbshipit-source-id: 787426928fed59c3e7e9182a4c64ef1aa72d6c4f