Make erlang_application() macro compatible with select()
Summary:
# Context
The `erlang_app()` rule is wrapped in an `erlang_application()` macro that does some normalizations, mostly so we can have short names for OTP deps. In particular, this means applying normalization functions to attrs such as `applications`

# Problem
As it is, it is not possible to use `select({...})` expressions in attrs such as `applications`, as configuration resolution happens after the macros run, and iterating over a select-expression fails.

# This diff
Buck2 provides the `select_map()` function to handle this scenario, so we use that instead.

# NB.
Unfortinately we are not able to do something similar with `erlang_tests()`, since it is also "filtering out"  elements from the `deps` attr, and there is afaics no function provided by byck2 to do this.

Reviewed By: TheGeorge

Differential Revision: D78158825

fbshipit-source-id: 53daec168235550e066408ac703e82f56e04f32b