Skip processed modules in manifest-module
Summary:
For each package path, the function loops and adds all parent directories to the module dictionary. The diff adds functionality to break out of the loop if it has already added a parent directory from a previous package path.

This diff reduced the time for this step from approximately 4 seconds to 1-2 seconds when running large targets, locally on my OD.

Reviewed By: zsol

Differential Revision: D78112700

fbshipit-source-id: a2a195e614d2a9228a1dd8101c4f8cf0a79b0996