Remove internal multi-toolchain logic
Summary:
# Context
The Erlang rules current support "multi-version erlang toolchains", that is, a single buck2 toolchain contains several erlang versions, and targets can be built with one or more of those

# Problem
* This was requested but never used, and there is no appetite for using it
* It is adding a lot of internal complexity, making it harder to introduce further optimizations, such as using depsets
* It leads to very high memory usage even when one version is used to build all targets
* We are moving to using buck2 configurations for toolchain resolution and this feature makes it rather awkward to implement

# This diff
As a first step, we remove all the internal logic dealing with multi-version toolchains. Essentially, all deps were previously indexed by the actual toolchain to use, and we are now removing this.

At this point we don't change the interface of the toolchain configuration: it still receives a primary toolchain and secondary toolchains that are currently ignored. This cleanup will happen  later in the stack

Reviewed By: TD5, michalmuskala

Differential Revision: D78237132

fbshipit-source-id: 8355e28412b810bbf88a085b0d92c320338cafd5