Remove secondaries from multi_version_toolchain_rule
Summary:
# Context
The `multi_version_toolchain_rule()` was used to define Erlang toolchains with more than one version of OTP, so that releases could then be built with more than one version.

We are in the process of removing support for this feature. In particular, in D78237132, we made it so that the internal logic of the rules ignore any toolchain other than the "primary" one (the first one in the list of targets given to the `multi_version_toolchain_rule`)

# Problem
We still expect to receive a list of toolchains, even though only one is used.

# This diff
Change the rule so that it receives a single toolchain. This way, we remove all places in the code that assume that more than one toolchain was used.

At this point, the rule will be a dummy wrapper. We will eliminate completely next

Reviewed By: TD5, michalmuskala

Differential Revision: D78273664

fbshipit-source-id: 56b7d19791253736c60487537aa0b314d6a07e0a