Remove the multi_version_toolchain_rule
Summary:
# Context
We are in the process of removing the support for multi-version toolchains, as it adds internal complexity and was never really used.

# This diff
After the changes in the stack, the rule is just a wrapper around a `ErlangToolchainInfo`. So here we remove the tool and promote `erlang_toolchain()` to toolchain-rule. In the process, we get rid of usages of the "primary" word, as this only made sense in the multi-version setting.

Reviewed By: michalmuskala

Differential Revision: D78280581

fbshipit-source-id: 14d5f6765457b7356ea0f1d30b2b6421f9f1c26d