erlang: tidy dependency_finalyzer
Summary:
Because of how we sometimes converted modules to be in the form `foo.beam` and sometimes not - just used `foo.erl` (specifically in `dependency_analyzer`); it meant that dependency collection in `dependency_finalyzer` was effectively recursive only for headers - for modules the `.erl` and `.beam` extensions wouldn't match and we'd never recurse.

This change makes this behaviour explicit and always uses just the file basename as the key. There's no functional change in this diff.

Reviewed By: TD5

Differential Revision: D78332107

fbshipit-source-id: 3250b9030792d373d5cbe6b94fda5216e631080e