erlang: merge deps files
Summary:
For each application, instead of exporting a map of `filename => dep file`, we merge the dep files into a single file with content of `filename => dep file content`.

This is done to help with buck2 memory - keeping track of all the edges from finalyzer to individual dep files was consuming large amount of memory.

This shouldn't impact incrementality - with the only change that we might need to re-run the merge step for a particular application - but we don't change what we depend on - before we depended on N files, now we depend on one file that includes the contets of all those files inside of it.

Reviewed By: TD5

Differential Revision: D78340846

fbshipit-source-id: f71db379aa26e9c1961620e2a6200493785e7514