erlang: separate storage for header & beam deps
Summary:
We separate storage for header dep files and beam dep files.

In particular, for running the `dependency_finalyzer` script, the only module dep file we need, is the module we're analysing - not all the other modules, since we only do recursive analysis for headers.

This means we don't even need to store the resulting dep files for modules beyond the compilation of the application itself.

This significantly reduces the number of files & modules we need to keep track of as input to the finalyzer script, speeding up the builds and reducing buck memory use.

This should also help with incrementality - since we severed connection between unrelated beam files on the buck side.

The dep files were also used for detection of module conflicts - instead of them, we leverage the beam file mapping to achieve the same goal.

Reviewed By: TD5

Differential Revision: D78334086

fbshipit-source-id: 1651f25129354588773b466a153c45590888eed1