Make annotation_processor_deps an exec_dep()
Summary:
This param is mostly unused now, because we tend to use the `plugins` param (ideally, we'd finish that migration off and not have this at all, I think, but whatever, people have better things to do).

This diff changes is so that when it is used, the deps are configured for the execution platform. That makes sense, because they're going to be used at _build_ time, i.e. on the execution platform.

Reviewed By: NavidQar

Differential Revision: D78558896

fbshipit-source-id: fff7f8fbb53ba631b40bf15ec548660206e5cd55