Introduce non_exec_dep_plugins_deprecated param
Summary: I want `plugins` to be configured to build for the execution platform, since that is where they are going to run. However, there is some databinding code that doesn't work with this at all, so providing an opt-out for them as I don't want to have to go and make it work, since it is old and unowned and quite the mess.

Reviewed By: NavidQar

Differential Revision: D78560599

fbshipit-source-id: 6e40159a6703ff528a514b26c5705704c76e46a8