Make kotlin_compiler_plugins take attrs.dep() instead of attrs.source()
Summary:
In the next diff, we want to make it so that `kotlin_compiler_plugins` are configured for the execution platform, since that is where they are run.

There is no version of `attrs.source()` that is configured for the execution platform, so instead we have to use `attrs.exec_dep()`. This diff does the intermediate step of moving to `attrs.dep()`, since we also have to change the param to be a list of tuples rather than a dict, because "Dependency" is not hashable. This list of tuples mirrors what we do for `plugins`.

Reviewed By: NavidQar

Differential Revision: D78734785

fbshipit-source-id: 8bb82c957b034ff4f2fe86c36e39ada44acc4d44