Avoid throwing exception on Clang failure in dist_lto_opt_darwin
Summary:
Previously this script made a clang invocation using `subprocess.check_call` which throws an exception if the invoked command fails. The stack trace for this exception clutters build failure logs, hiding the true cause of the issue -- the failed clang invocation.

Instead avoid throwing an exception if clang fails, just print stderr and exit with a non-zero exit code.

Reviewed By: ellishg

Differential Revision: D78912025

fbshipit-source-id: 4f4057d26501b82edd33409ea60b8fdfba74f36f