Replace map snapshot representation by a list
Summary:
## Context
Enabling KSIC on ig4a resulted in approximately 18K additional action executions for ig4a. This change also caused a memory regression of over 10%.

Both issues are very likely due to the addition of a snapshot action for every incremental target, which explains the observed increase in executions and memory usage.

## This diff
In a previous diff (D78347915), we enabled projection for classpath snapshot with the goal of reducing notes in the build graph. However, the JSON projection cannot treat each entry as a key-value pair. Due to this limitation, we had  to parse a list of key-value strings instead.

This diff simplifies the logic by using a list representation for snapshots instead of a map.

Reviewed By: IanChilds

Differential Revision: D78885024

fbshipit-source-id: c34442b2f1dbd325e971a0e3f2813eebd350dcee