Prepare JavaCD and KotlinCD for content-based inputs
Summary:
The inline comment explains what we're doing here: we have a "placeholder" version of the argfile, that we use when running dep-files, and the real argfile that we ignore when running dep-files but use for everything else.

This doesn't turn on content-based paths anywhere, but it allows us to start turning them on for inputs to these actions.

Reviewed By: NavidQar

Differential Revision: D78665894

fbshipit-source-id: 91b574dccd2a51759ce21d9f4e3aad4e502e8f90