Allow windows_resource() to take deps
Summary:
We have a .rc file which does an include of a file in a different directory that is not part of the same package.  In order to support this, allow `windows_resource()` to accept a `deps` argument, and merge this in with the existing preprocessor info.

Also, less critical, but I decided to set the `.res` file as the rule's `default_output`.  It's nice to be able to `buck2 build` the resources target directly for debugging purposes.

X-link: https://github.com/facebook/buck2/pull/1015

Reviewed By: scottcao

Differential Revision: D77395863

Pulled By: JakobDegen

fbshipit-source-id: 107348431f63a6f28e6b7ad17e6f1fc488d0a3e6