patch ctypes.util.find_library to work with PARs
Summary:
`find_library` in OSS does a commendable job of calling `ldconfig` and `gcc` and `ld` to try and find DSOs to be passed to `dlopen`, but unfortunately none of the above work in our pars because by the time `find_library` is invoked, `LD_LIBRARY_PATH` gets cleared out.

This diff patches the function to apply heuristics that work better in our environment (and also don't call build tools) for macos and linux.

We still fall back to the original implementation if our heuristics fail.

Reviewed By: AishwaryaSivaraman

Differential Revision: D79720303

fbshipit-source-id: 1942dae8faba316ed51f37d9eb7b7891ff9d0117