buck2: remote_python_toolchain
Summary:
X-link: https://github.com/facebook/buck2/pull/1060

This diff introduces a new `remote_python_toolchain` that downloads a CPython distribution from GitHub and uses that for `python_*` rules instead of what's available on the system. It also uses that by default from the demo toolchain that's used in examples.

Reviewed By: podtserkovskiy

Differential Revision: D79552057

fbshipit-source-id: ed5b5f4aa714292ab82a801107c040a1f18cda24