Pass whole path to /WHOLEARCHIVE?
Summary:
While testing D78969509, we found that some links failed due to `/WHOLEARCHIVE` only receiving the basename of a lib and not the whole path. It looks like this was intentional when implemented by D24921551. However, in my local testing I was not able to reproduce a case where passing full path to `/WHOLEARCHIVE` created a build failure. So throwing this diff up to try and simplify our linker command.

We use `link_whole = True` for rust cxx bridge, and the linker issue was reproing on arvr win modes as they weren't using the ld_win.py wrapper

```
  = note: lld-link: error: could not open 'libbuiltin_executors_rust-bridge.lib': no such file or directory␍
          lld-link: error: could not open 'lib_iobuf-bridge.lib': no such file or directory␍
```

#buildall

Reviewed By: dtolnay

Differential Revision: D79901485

fbshipit-source-id: 5288efdb1d4c6b5fea293359a8a133230d135919