Fix regression caused by extra layer of bxl hash.
Summary:
This was previously fixed in cb62fee for fbsource, with a comment indicating that we would wait and see if it also needs to be fixed for non-fbsource.  I ran into this today, so fixing it for all users.

X-link: https://github.com/facebook/buck2/pull/1070

Reviewed By: IanChilds

Differential Revision: D79911649

Pulled By: JakobDegen

fbshipit-source-id: bc39b4011be1e227d2a55afff732224a71a8c5b0