Factor out common functionality
Summary:
The same logic with the same purpose existed in multiple places, so we factor it out into a utility module. Thanks to TheGeorge for giving the the nudge to do this.

We don't do this for `boot_script_builder`, because it executes before the utility module is in place (specifically, I think the symlink is added afterwards).

Reviewed By: TheGeorge

Differential Revision: D79717803

fbshipit-source-id: 4384b46ecf68349b48f229c95dad54b5349c687c