Codesigning: simplify codesign manifest tree
Summary:
**Problem**

The codesign manifest tree is a relatively complex output - a tree of JSON files, containing signing commands. This means that all clients need to understand the correct ordering of processing codesigning commands (inside out) and implement the required parsing.

 The goal is to make misusing the codesigning manifest much harder and simplify clients.

**Solution**

Postprocess the output and produce a simple list of codesigning commands. This flattens the nested structure of JSON files and provides an explicit ordering of the codesigning commands, removing the need for clients to implement this logic.

Reviewed By: manicaesar

Differential Revision: D79993153

fbshipit-source-id: ce5e3a06353cdf34e452c3064197e725d6c2de9c